Add support for long-lived tokens
A new authentication model was introduced in version `0.77` of Home
Assistant. The underlying node module that powers this Hubot package has
been updated to support that new model, with the legacy API password
being deprecated in a future release of Home Assistant.

Fixes #8.

NOTE: This will require a configuration change for installed users.
Those not wanting to upgrade quite yet should remain on `v1.0.2`.